Add pid testing for entire chassis
There's some weird memory errors that I should clean up before we try to
do stuff with this.

+++ base.c
@@ -30,14 +30,14 @@ void printEncoders() {
     printf("\n");
 }
 
-void drive(float x, float y) {
+void drive(double x, double y) {
     setMotor(top, x);
     setMotor(bottom, -x);
     setMotor(left, y);
     setMotor(right, -y);
 }
 
-void spin(float speed) {
+void spin(double speed) {
     setMotor(top, speed);
     setMotor(bottom, speed);
     setMotor(left, speed);
