+++ motor.h
@@ -13,7 +13,7 @@ typedef struct Motor Motor;
 Motor* createMotor(int m1, int m2, int pwm, int encA, int encB);
 
 // Speed is in range [-1,1]
-void setMotor(Motor* m, float speed);
+void setMotor(Motor* m, double speed);
 
 void stopMotor(Motor* m);
